refactor: result-list into smaller components
